% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_02_export_r_data.R
\name{export_ndata_clean}
\alias{export_ndata_clean}
\title{Exports the \code{ntits_clean} dataset to the "data/" folder}
\usage{
export_ndata_clean(
  my_tdata = here::here("output", "tables", "ndata_parcond.csv")
)
}
\arguments{
\item{my_tdata}{The path to "ndata_parcond.csv" because, internally, this function calls the}
}
\value{
The path to the dataset generated
by \code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}} because, internally, this function calls
\code{\link[ppl.tits:tdata_upD_final]{tdata_upD_final}} and thus requires the same input dataset.
}
\description{
FOR INTERNAL USE ONLY. DO NOT RUN.
}
\examples{
\dontrun{
# DO NOT USE!
export_ndata_clean()
}
}
