% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_02_export_r_data.R
\name{export_nestling_aggreg}
\alias{export_nestling_aggreg}
\title{Export the 'tits nestling' aggregated dataset (i.e. \code{tits_nestling_data.csv})}
\usage{
export_nestling_aggreg(myrawdata)
}
\arguments{
\item{myrawdata}{The path to the raw tits data object (\code{raw_tits_data.rda}). Cf.
\code{\link[ppl.tits:import_raw_tits_data]{import_raw_tits_data}}.}
}
\value{
The path to the aggregated nestling tits dataset: \code{tits_nestling_data.csv}.
}
\description{
(THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!) \cr
The \code{export_nestling_aggreg()} function aggregates the raw tits data to create a table
summarizing tits' breeding success for each nestbox each year where breeding occurred. As such,
it only keeps information on nestlings/juveniles and disregards the lines about adult tits. Then, the
function exports the created dataset (.csv) and returns its path. \strong{Actually}, the function
also creates the folders in which the created dataset will be stored.
}
\examples{
\dontrun{
export_nestling_aggreg(myrawdata = raw_tits)
}
}
