% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_00_import_raw_data.R
\name{import_raw_tits_data}
\alias{import_raw_tits_data}
\title{Import the raw tits dataset}
\usage{
import_raw_tits_data(mypath = here::here("mydata", "ppl_dijon_tits_data.csv"))
}
\arguments{
\item{mypath}{The absolute path to the raw .csv file.}
}
\value{
A tibble (i.e. a kind of improved data.frame). For further information on tibbles, please refer to
the \code{tidyverse} or \link[readr]{readr} documentation.
}
\description{
Imports ONLY the raw dataset for the "tits" part of the PubPrivLands project, and drops
lines with no information on 'success' as well as one useless column. \cr To avoid errors,
if new data are added using this function, they should be formatted according to the first table
I tailored this function for (i.e. same columns, no special characters, no spaces, etc.; the only
tolerated differences may be different rows and cell values)!
}
\examples{
\dontrun{
mydata <- import_raw_tits_data(mypath = "C:/path/to/my/data.csv") # This example cannot work,
# it needs an existing path!
}
}
