% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntits_clean}
\alias{ntits_clean}
\title{Tits nestling data (cleaned version)}
\format{
\subsection{\code{ntits_clean}}{

A data frame (tibble) with 406 rows and 47 columns:
\describe{
\item{country}{Country name}
\item{iso2, iso3}{2 & 3 letter ISO country codes}
\item{year}{Year}
...
}
}
}
\source{
\url{https://github.com/mrelnoob/ppl.tits}
}
\usage{
ntits_clean
}
\description{
The cleaned dataset showing the
}
\keyword{datasets}
