% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{raw_tits_data}
\alias{raw_tits_data}
\title{Raw tits dataset}
\format{
\subsection{\code{raw_tits_data}}{

A data frame (tibble) with MANY ROWS AND COLUMNS§§§ TO BE CONTINUED
\describe{
\item{\strong{id_nestbox}}{Categorical -- The nestbox ID.}
\item{\strong{xxx}}{TO BE CONTINUED UPON PUBLICATION***...}
...
}
}
}
\source{
\url{https://github.com/mrelnoob/ppl.tits}
}
\usage{
raw_tits_data
}
\description{
This is the raw data used for the first part of the "tits reproduction study" of the PubPrivLands
project containing observations of both Great tits (\emph{Parus major}) and Blue tits (\emph{Cyanistes
caeruleus}) adults and nestlings in Dijon between 2019 and 2022. \cr
This dataset was imported and corrected by the
\code{\link[ppl.tits:import_raw_tits_data]{import_raw_tits_data}} function.
}
\details{
For further details regarding data acquisition, processing or metadata, readers are referred to: i) the
documentation of the \code{{ppl.tits}} package functions and related reports (cf. the package's
\href{https://github.com/mrelnoob/ppl.tits}{README file}; ii) the associated article's Materials & Methods
section; and iii) the article's Supplementary Materials (LINKS?***).
}
\keyword{datasets}
