% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_02_export_r_data.R
\name{read_from_path}
\alias{read_from_path}
\title{Read stored RData objects from their path}
\usage{
read_from_path(mypath)
}
\arguments{
\item{mypath}{Path to the name of the stored data (created by
\code{\link[base:save]{save}}).}
}
\value{
The stored data (exactly like \code{pkg::stored_data} would).
}
\description{
(THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!).
}
\examples{
\dontrun{
# read_from_path(mypath = path_to_package_data)
}
}
