% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{tdata_upD_final}
\alias{tdata_upD_final}
\title{Update and export the tits dataset completed with imputed missing values}
\usage{
tdata_upD_final(my_tdata = here::here("output", "tables", "ndata_parcond.csv"))
}
\arguments{
\item{my_tdata}{The path to the dataset generated
by \code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}. Note that this argument is not
necessarily required but is strongly advised.
Therefore, if you want to reproduce this work, you should not call this function unless you
have previously built the {targets} pipeline described in _targets.R (e.g. by using
targets::tar_make())! If you DO NOT KNOW what I'm talking about, then you should
read the README file (https://github.com/mrelnoob/ppl.tits) and PROBABLY NOT modify the
function's arguments (see examples section).}
}
\value{
A list containing five elements: 1) the updated "clean" dataset (a tibble accessible with
\code{tdata_upD_final()$clean_dataset}); 2) the dataset before imputations but containing all data
corrections applied by the function (a tibble accessible with
\code{tdata_upD_final()$priorimp_dataset}); 3 and 4) tables showing the out-of-bag imputation error
for both species and for each of the four "buffer radius" (accessible with \code{tdata_upD_final()$impute_error}
followed by the desired species code (e.g. "_pm"; cf. examples)); and 5) the path to the exported
.csv file of the updated "clean" dataset (\code{ndata_clean.csv}; accessible with \code{tdata_upD_final()$path}).
}
\description{
(THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!) \cr
The function is the \strong{fourth} and last of a series of functions meant to
update and complete the tits dataset with its independent variables (IV - i.e. predictors and
covariates).
The \code{tdata_upD_final} function loads and modifies the dataset generated by the previous
update function (\code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}) by correcting a few
last errors in the data and by imputing missing values using \strong{Random Forest} regressions. For
consistency, as the global dataset contains IV extracted from four sizes of buffer around nestboxes,
missing data imputations have been performed on each sub-dataset (e.g. 100m buffer-based data,
150m buffer-based data, etc.), independently. \cr
For more details regarding the imputation process and associated errors, please refer to the
\emph{Exploratory data analyses and preparation report}.
When it's done processing, the function exports the updated dataset
(\code{ndata_clean}) as a new .csv file.
}
\examples{
\dontrun{
# To obtain the data and imputation error regarding the 50m buffer data:
mydata <- tdata_upD_final()$clean_dataset
mydata \%>\% dplyr::filter(dist == 50) -> mydata_50
mydata <- tdata_upD_final()$priorimp_dataset # May be useful to compare the results of missing
# data imputation.
myerror_pm <- tdata_upD_final()$impute_error_pm
myerror_cc <- tdata_upD_final()$impute_error_cc
path_to_csv <- tdata_upD_final()$path
}
}
