% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{tdata_upD_rawiv}
\alias{tdata_upD_rawiv}
\title{Update and export the tits dataset with all its "raw" independent variables (IVs)}
\usage{
tdata_upD_rawiv(
  my_tdata = here::here("output", "tables", "ndata_temp.csv"),
  my_iv_data = here::here("mydata", "tits_predictors.csv")
)
}
\arguments{
\item{my_tdata}{The name of the \strong{target} that contains the dataset generated
by \code{\link[ppl.tits:tdata_upD_temp]{tdata_upD_temp}}. Note that the function
\strong{will not work} unless the name of the appropriate target is correctly specified.
Therefore, if you want to reproduce this work, you cannot call this function unless you
have previously built the {targets} pipeline described in _targets.R (e.g. by using the
make.R file of the project)! If you DO NOT KNOW what I'm talking about, then you should
read the README file (https://github.com/mrelnoob/ppl.tits) and NOT use this function!}

\item{my_iv_data}{The dataset (.csv) generated by PostGIS (cf. ppl_log.docx) and containing the
"raw" independent variables (IV = predictors and covariates) meant to describe the local
environment around each nestbox (= zonal statistics). It is called "raw" as opposed to the
more "advanced" IV that will be computed by the next functions to generate the "final" tits
dataset.}
}
\value{
A list containing 1) the updated dataset (a tibble accessible with
\code{tdata_upD_rawiv()$dataset}), and 2) the path to the exported .csv file of the updated
dataset (accessible with \code{tdata_upD_rawiv()$path}).
}
\description{
The function is the \strong{second} of a series of functions meant to update and
complete the tits dataset with its independent variables (i.e. predictors and covariates).
The \code{tdata_upD_rawiv} function loads and modifies the dataset generated by the previous
update function (\code{\link[ppl.tits:tdata_upD_temp]{tdata_upD_temp}}) in several ways:
\itemize{
\item First, it performs a left join between both input datasets (i.e. \code{my_tdata} and
\code{my_iv_data} - see below).
\item Second, it computes two new variables: namely \emph{woody_area} and \emph{open_area} from
already existing columns.
\item Third, it reorganises the dataset for improved clarity.
When it's done processing, the function exports the updated dataset as a new .csv file.
}
}
\examples{
\dontrun{
mydata <- tdata_upD_rawiv()$dataset
path_to_csv <- tdata_upD_rawiv()$path
}
}
