% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{tdata_update_temp}
\alias{tdata_update_temp}
\title{Update and export the tits dataset with temperature related variables}
\usage{
tdata_update_temp(
  myboxtemp_data = here::here("mydata", "paired_boxtemp.csv"),
  mytits_data = here::here("output", "tables", "tits_nestling_data.csv")
)
}
\arguments{
\item{myboxtemp_data}{The nestbox-temperature stations pairing dataset (.csv).}

\item{mytits_data}{The tits nestling aggregated dataset (.csv). Cf.
\code{\link[ppl.tits:export_nestling_aggreg]{export_nestling_aggreg}}.}
}
\value{
A tibble with an improved version of the tits dataset.
}
\description{
The function is the \strong{first} of a series of functions meant to update and complete
the tits dataset with its \emph{independent variables} (i.e. predictors and covariates).
The \code{tdata_update_temp} function modifies the \emph{aggregated tits dataset} in several ways:
\itemize{
\item First, it assigns to each observation (nestbox for a given year) the temperature station it is
paired with (giving the new column "\emph{temp_station_id}").
\item Second, it reorganizes the date columns (deleting 2 and updating the others).
\item Third, it creates a new \code{factor} variable called "\emph{breeding window}" that indicates whether
the reproduction event occurred at the beginning or at the end of the breeding season.
\item Fourth, it computes five \strong{temperature related} variables. These variables are
\emph{cumdd_30} (the cumulative day-degrees for the 30 days prior to the laying date);
\emph{min_t_before} and \emph{min_t_between} (which are the minimum temperature recorded during the
30 days prior to the laying date and between the laying and the flight date, respectively);
\emph{mean_winter_t} and \emph{sd_winter_t}) (which are the mean recorded temperature during the
four month of winter across 2019-2022 and its standard deviation, respectively).
}
}
\examples{
\dontrun{
mydata <- tdata_update_temp()
}
}
