% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_02_exploratory_data_analyses.R
\name{uni.boxplots}
\alias{uni.boxplots}
\title{Univariate boxplots}
\usage{
uni.boxplots(
  dataset,
  MAR = c(0.5, 4.1, 1.1, 1.5),
  CEX.LAB = 1,
  FONT.LAB = 2,
  BTY = "n",
  FG = "gray35",
  COL.AXIS = "gray35",
  COL.LAB = "gray20",
  CEX.PAR = 0.8,
  TCL = -0.3,
  MGP = c(2.4, 0.6, 0),
  OMA = c(1, 0, 0, 0),
  TYPE = "n",
  BORDER = "lightcoral",
  COL = "moccasin",
  LTY = 1,
  STAPLEWEX = 0,
  WHISKLWD = 2,
  BOXWEX = 0.7,
  BOXLWD = 0.1,
  MEDLWD = 2.6,
  PCH = 19,
  ...
)
}
\arguments{
\item{dataset}{The input dataset containing all variables to be plotted. It may contain all kinds of
variables, the \code{uni.boxplots} function will automatically detect and plot numeric variables (columns).}

\item{MAR}{A numerical vector of the form \code{c(bottom, left, top, right)} which gives the number of lines
of margin to be specified on the four sides of the plot. The default is \code{c(0.5,4.1,1.1,1.5)}.}

\item{CEX.LAB}{The magnification to be used for x and y labels relative to the current setting of \code{cex}.}

\item{FONT.LAB}{The font to be used for x and y labels.}

\item{BTY}{A character string which determined the type of box which is drawn about plots. If \code{BTY} is
one of "o", "l", "7", "c", "u", or "]" the resulting box resembles the corresponding upper
case letter. A value of "n" suppresses the box (the default).}

\item{FG}{The color to be used for the foreground of plots. This is the default color used for things
like axes and boxes around plots (defaults to "gray35").}

\item{COL.AXIS}{The color to be used for axis annotation. Defaults to "gray35".}

\item{COL.LAB}{The color to be used for x and y labels. Defaults to "gray20".}

\item{CEX.PAR}{A numerical value giving the amount by which plotting text and symbols should be
magnified relative to the default (for \code{par}, the panel manager). This starts as 1 when a device
is opened, and is reset when the layout is changed, e.g. by setting \code{mfrow}. Defaults to 0.8.}

\item{TCL}{The length of tick marks as a fraction of the height of a line of text. The default
value is -0.3.}

\item{MGP}{The margin line (in \code{mex} units) for the axis title, axis labels and axis line.
Note that \code{mgp[1]} affects title whereas \code{mgp[2:3]} affect axis. The default is c(2.4, 0.6, 0).}

\item{OMA}{A vector of the form \code{c(bottom, left, top, right)} giving the size of the outer margins
in lines of text.}

\item{TYPE}{The type of boxplot to draw. Default is "n".}

\item{BORDER}{An optional vector of colors for the outlines of the boxplots. The values in border
are recycled if the length of border is less than the number of plots. Default is "lightcoral".}

\item{COL}{If col is non-null it is assumed to contain colors to be used to colour the bodies of
the boxplots. Default is "moccasin".}

\item{LTY}{The line type. Line types can either be specified as an integer (0=blank, 1=solid (default),
2=dashed, 3=dotted, 4=dotdash, 5=longdash, 6=twodash) or as one of the character strings "blank",
"solid", "dashed", "dotted", "dotdash", "longdash", or "twodash", where "blank" uses ‘invisible lines’
(i.e., does not draw them).}

\item{STAPLEWEX}{Staple line width expansion, proportional to box width. Default is 0.}

\item{WHISKLWD}{Whisker line width expansion. Default is 2.}

\item{BOXWEX}{A scale factor to be applied to all boxes. When there are only a few groups, the
appearance of the plot can be improved by making the boxes narrower. Default is 0.7.}

\item{BOXLWD}{Width of boxplot outer lines. Default is 0.1.}

\item{MEDLWD}{Width of the median line. Default is 2.6.}

\item{PCH}{The type of points to be drawn for outliers. Default is 19. See \code{\link[graphics:points]{points}}
for possible values and their interpretation.}

\item{...}{Any other parameter that can be incorporated in \code{\link[graphics:boxplot]{boxplot}}.}
}
\value{
A panel of univariate boxplots.
}
\description{
The \code{uni.boxplots} function draws, within a single panel, an independent boxplot for each
numeric (continuous or discrete) variable in a given dataset. It is particularly useful for data
exploration (e.g. Zuur \emph{et al.}, 2010). For instance, to simultaneously observe the
distributions of all numeric variables or \strong{to detect their univariate outliers}.
}
\details{
The \code{uni.boxplots} function only modifies the graphical parameters of the
\code{\link[graphics:boxplot]{boxplot}} function in the \code{graphics} package to match some predefined
preferences. Therefore, default values of \code{uni.boxplots} create nice looking boxplots but retain
default \emph{heuristic} aspects of \code{boxplot} (such as the length of whiskers or the plotting of
outliers). These aspects can however be changed as in \code{\link[graphics:boxplot]{boxplot}}. \cr
On the other hand, panel parameters are internally controlled using \code{par}. However, to avoid unforeseen
conflicts with other internal parameters, it is not possible to tune panel parameters as we would
do with \code{par}. Instead, parametrization is only possible with the given subset of parameters.
}
\note{
To avoid \emph{recursive argument errors}, internal arguments should be called using upper case
letters (e.g. CEX.LAB = 0.9) whereas other arguments from the \code{boxplot} function should be called with
their normal case writing (e.g. outline = FALSE)!
}
\examples{
data("mtcars")
uni.boxplots(dataset = mtcars)
}
