% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_02_exploratory_data_analyses.R
\name{uni.histograms}
\alias{uni.histograms}
\title{Histogram Panel}
\usage{
uni.histograms(
  dataset,
  MAR = c(3, 2, 0.5, 1.5),
  CEX.LAB = 1.2,
  FONT.LAB = 2,
  BTY = "n",
  FG = "gray35",
  COL.AXIS = "gray35",
  COL.LAB = "gray20",
  CEX.PAR = 0.6,
  TCL = -0.3,
  MGP = c(1.7, 0.6, 0.1),
  OMA = c(1, 0, 1, 0),
  LAB = c(5, 10, 7),
  BREAKS = 10,
  COL = "moccasin",
  BORDER = "white"
)
}
\arguments{
\item{dataset}{The input dataset containing all variables to be plotted (must be a \code{data.frame} with
at least 2 variables). It may contain all kinds of columns, the \code{uni.histograms} function will
automatically detect and plot numeric variables (columns).}

\item{MAR}{A numerical vector of the form \code{c(bottom, left, top, right)} which gives the number of lines
of margin to be specified on the four sides of the plot. The default is \code{c(0.5,4.1,1.1,1.5)}.}

\item{CEX.LAB}{The magnification to be used for x and y labels relative to the current setting
of \code{CEX.PAR}.}

\item{FONT.LAB}{The font to be used for x and y labels.}

\item{BTY}{A character string which determined the type of box which is drawn about plots. If \code{BTY} is
one of "o", "l", "7", "c", "u", or "]" the resulting box resembles the corresponding upper
case letter. A value of "n" suppresses the box (the default).}

\item{FG}{The color to be used for the foreground of plots. This is the default color used for things
like axes and boxes around plots (defaults to "gray35").}

\item{COL.AXIS}{The color to be used for axis annotation. Defaults to "gray35".}

\item{COL.LAB}{The color to be used for x and y labels. Defaults to "gray20".}

\item{CEX.PAR}{A numerical value giving the amount by which plotting text and symbols should be
magnified relative to the default (for \code{par}, the panel manager). This starts as 1 when a device
is opened, and is reset when the layout is changed, e.g. by setting \code{mfrow}. Defaults to 0.8.}

\item{TCL}{The length of tick marks as a fraction of the height of a line of text. The default
value is -0.3.}

\item{MGP}{The margin line (in \code{mex} units) for the axis title, axis labels and axis line.
Note that \code{mgp[1]} affects title whereas \code{mgp[2:3]} affect axis. The default is c(2.4, 0.6, 0).}

\item{OMA}{A vector of the form \code{c(bottom, left, top, right)} giving the size of the outer margins
in lines of text.}

\item{LAB}{A numerical vector of the form \code{c(x, y, len)} which modifies the default way that axes
are annotated. The values of \code{x} and \code{y} give the (approximate) number of tickmarks on the x and y
axes and len specifies the label length. The default is \code{c(5, 5, 7)}. Note that this only affects
the way the parameters xaxp and yaxp are set when the user coordinate system is set up, and is
not consulted when axes are drawn. \code{len} is unimplemented in \code{R}.}

\item{BREAKS}{One of:
\itemize{
\item a vector giving the breakpoints between histogram cells,
\item a function to compute the vector of breakpoints,
\item a single number giving the number of cells for the histogram,
\item a character string naming an algorithm to compute the number of cells (see
\code{\link[graphics:hist]{hist}}),
\item a function to compute the number of cells.
}

In the last three cases the number is a suggestion only; as the breakpoints will be set to pretty
values, the number is limited to 1e6 (with a warning if it was larger). If breaks is a function,
the x vector is supplied to it as the only argument (and the number of breaks is only limited by
the amount of available memory).}

\item{COL}{The color of the bar of the histograms (bins).}

\item{BORDER}{The color of the border of the bars.}
}
\value{
A panel of histograms.
}
\description{
The \code{uni.histograms} function draws, within a single panel, an independent histogram
or each numeric (continuous or discrete) variable in a given dataset. It is particularly useful
for data exploration (e.g. Zuur \emph{et al.}, 2010). For instance, to simultaneously observe
the distributions of all numeric variables and determine which one will require transformation.
}
\examples{
uni.histograms(dataset = iris[,1:4])
}
