% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_02_export_r_data.R
\name{export_ndata_clean}
\alias{export_ndata_clean}
\title{Exports the \code{ntits_clean} dataset to the "data/" folder}
\usage{
export_ndata_clean(
  my_clean_data = here::here("output", "tables", "ndata_clean.csv")
)
}
\arguments{
\item{my_clean_data}{The path to "ndata_clean.csv".}
}
\value{
The path to "data/ndata_clean.rda".
}
\description{
FOR INTERNAL USE ONLY. DO NOT RUN.
}
\examples{
\dontrun{
# DO NOT USE!
export_ndata_clean()
}
}
