% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R, R/01_02_export_r_data.R
\name{export_nestling_aggreg}
\alias{export_nestling_aggreg}
\title{Export tits nestling aggregated dataset}
\usage{
export_nestling_aggreg(myrawdata)

export_nestling_aggreg(myrawdata)
}
\arguments{
\item{myrawdata}{The path to the raw tits data object (\code{raw_tits_data.rda}). Cf.
\code{\link[ppl.tits:import_raw_tits_data]{import_raw_tits_data}}.}
}
\value{
A CSV table and its path.

The path to the aggregated nestling tits dataset: \code{tits_nestling_data.csv}.
}
\description{
The \code{export_nestling_aggreg} function exports the dataset (.csv) created by the
\code{aggreg_by_nest} function, that is the tits nestling data aggregated by nestbox and year. The
input dataset should therefore be the \strong{raw dataset}, not an object created by the
aggregating function!

(THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!) \cr
The \code{export_nestling_aggreg()} function aggregates the raw tits data to create a table
summarizing tits' breeding success for each nestbox each year where breeding occurred. As such,
it only keeps information on nestlings/juveniles and disregards the lines about adult tits. Then, the
function exports the created dataset (.csv) and returns its path. \strong{Actually}, the function
also creates the folders in which the created dataset will be stored.
}
\examples{
\dontrun{
export_nestling_aggreg(myrawdata = raw_tits)
}
\dontrun{
export_nestling_aggreg(myrawdata = raw_tits)
}
}
