% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{export_nestling_aggreg}
\alias{export_nestling_aggreg}
\title{Export tits nestling aggregated dataset}
\usage{
export_nestling_aggreg()
}
\value{
A CSV table and its path.
}
\description{
The \code{export_nestling_aggreg} function exports the dataset (.csv) created by the
\code{aggreg_by_nest} function, that is the tits nestling data aggregated by nestbox and year.
}
\examples{
\dontrun{
export_nestling_aggreg()
}
}
