% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_00_local_predictive_modelling.R
\name{local_quality_model}
\alias{local_quality_model}
\title{Build a predictive model of local habitat quality}
\usage{
local_quality_model(
  my_tdata = here::here("output", "tables", "ndata_final.csv")
)
}
\arguments{
\item{my_tdata}{The name of the \strong{target} that contains the dataset generated
by \code{\link[ppl.tits:tdata_upD_final]{tdata_upD_final}}. Note that this argument is not
necessarily required but is strongly advised.
Therefore, if you want to reproduce this work, you should not call this function unless you
have previously built the {targets} pipeline described in _targets.R (e.g. by using
targets::tar_make())! If you DO NOT KNOW what I'm talking about, then you should
read the README file (https://github.com/mrelnoob/ppl.tits) and PROBABLY NOT modify the
function's arguments (see examples section).}
}
\value{
This functions returns three things: 1) The Random Forest object built to be used for
prediction using new data (e.g. using
\code{\link[randomForest:predict.randomForest]{predict.randomForest}}; accessible through
\code{local_quality_model()$rf4pm}); 2) the OOB R2 stability plot (accessible through
\code{local_quality_model()$r_squared.stab}); and 3) the variables importance stability plot
(accessible through \code{local_quality_model()$var_importance.stab}).
}
\description{
The \code{local_quality_model} function builds a Random Forest (RF) regression model to
predict the local habitat quality for Great tits (\emph{Parus major} only) approximated by the species
clutch size. \cr
The RF model, fitted using the \code{\link[randomForest:randomForest]{randomForest}} package, is
build without hyperparameter tuning even though predictor variables importance is also computed.
Additionally, the \code{local_quality_model} function also assess the stability of the RF modelling
process by repeating it 100 times and reporting the associated out-of-bag R squared values as
well as permutation-based variable importance metrics (computed as averaged increase in Mean
Squared Error).
}
\note{
This function only produces outputs for the Great tits (\emph{Parus major}). Computations
for Blue tits (\emph{Cyanistes caeruleus}) were too unstable to be exploited due to insufficient
sample size.
}
\examples{
\dontrun{
# This line should work as long as "my_newdata" is an appropriately formatted dataset containing
# the same variables as the input RF model:
patch_quality <- predict.randomForest(
  ppl.tits::local_quality_model()$rf4pm,
  newdata = my_newdata)

ppl.tits::local_quality_model()$r_squared.stab # To generate the R2 stability plot
ppl.tits::local_quality_model()$var_importance.stab # To generate the variable importance
# stability plot
}
}
