% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntits_clean}
\alias{ntits_clean}
\title{Tits nestling data (cleaned version)}
\format{
\subsection{\code{ntits_clean}}{

A data frame (tibble) with 406 rows and 47 columns. Here's
\describe{
\item{\strong{id_nestbox}}{Categorical -- The nestbox ID.}
\item{\strong{site}}{Categorical -- The name of the site where the nestbox is located.}
\item{\strong{coord_x}}{Continuous -- The longitudinal coordinates of the nestbox location (Lambert-93
projection).}
\item{\strong{coord_y}}{Continuous -- The latitudinal coordinates of the nestbox location (Lambert-93
projection).}
\item{\strong{year}}{Categorical -- The year of observation (2019 to 2022).}
\item{\strong{species}}{Categorical -- The observed species (PM or CC).}
\item{\strong{xxx}}{TO BE CONTINUED UPON PUBLICATION***...}
...
}
}
}
\source{
\url{https://github.com/mrelnoob/ppl.tits}
}
\usage{
ntits_clean
}
\description{
This is the cleaned dataset from the first part of the "tits reproduction study" of the PubPrivLands
project containing observations for the Great tits (\emph{Parus major}) and Blue tits (\emph{Cyanistes
caeruleus}) nestlings aggregated at the nestbox level for each year of study. \cr
This dataset was generated by the first group of functions of the \code{{ppl.tits}} package that cleaned and
prepared all raw data linked to tits breeding or local environmental conditions. This dataset thus
contains the \strong{response} variables related to tits' \emph{reproduction success} and
\emph{physical condition}, as well as all potential \strong{explanatory} variables and \strong{covariates}
that could be used for analysis, with the exception of the \emph{connectivity} metrics that will be
computed by a second group of functions, giving the "final" tits dataset (\code{ndata_complete}; LINK when
ready?***). \cr
}
\details{
For further details regarding data acquisition, processing or metadata, readers are referred to: i) the
documentation of the \code{{ppl.tits}} package functions and related reports (cf. the package's
\href{https://github.com/mrelnoob/ppl.tits}{README file}; ii) the associated article's Materials & Methods
section; and iii) the articles's Supplementary Materials (LINKS?***).
}
\keyword{datasets}
