% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppl.tits-package.R
\docType{package}
\name{ppl.tits-package}
\alias{ppl.tits}
\alias{ppl.tits-package}
\title{ppl.tits: Contains all Data Analyses for the First Part Of Ppl.tits}
\description{
This package project is meant to clean and prepare the data as well as do all spatial and statistical analyses for the first part of the "tits" section of the PubPrivLands project.
}
\author{
\strong{Maintainer}: Francois-Marie Martin \email{francois-marie.martin@univ-fcomte.fr}

}
\keyword{internal}
