% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{tdata_upD_final}
\alias{tdata_upD_final}
\title{Update and export the tits dataset completed with imputed missing values}
\usage{
tdata_upD_final(my_tdata = here::here("output", "tables", "ndata_parcond.csv"))
}
\arguments{
\item{my_tdata}{The path to the dataset generated
by \code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}. Note that this argument is not
necessarily required but is strongly advised.
Therefore, if you want to reproduce this work, you should not call this function unless you
have previously built the {targets} pipeline described in _targets.R (e.g. by using
targets::tar_make())! If you DO NOT KNOW what I'm talking about, then you should
read the README file (https://github.com/mrelnoob/ppl.tits) and PROBABLY NOT modify the
function's arguments (see examples section).}
}
\value{
A list containing five elements: 1) the updated "clean" dataset (a tibble accessible with
\code{tdata_upD_final()$clean_dataset}); 2) the dataset before imputations but containing all data
corrections applied by the function (a tibble accessible with
\code{tdata_upD_final()$priorimp_dataset}); 3 and 4) tables showing the out-of-bag imputation error
for both species (accessible with \code{tdata_upD_final()$impute_error_pm} or
\code{tdata_upD_final()$impute_error_cc}); and 5) the path to the exported .csv file of the updated
"clean" dataset (\code{ndata_clean.csv}; accessible with \code{tdata_upD_final()$path}).
}
\description{
(THEORETICALLY MEANT FOR INTERNAL USE ONLY, use with caution!) \cr
The function is the \strong{fourth} and last of a series of functions meant to
update and complete the tits dataset with its independent variables (i.e. predictors and
covariates).
The \code{tdata_upD_final} function loads and modifies the dataset generated by the previous
update function (\code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}) by correcting a few
last errors in the data and by imputing missing values using \strong{Random Forest} regressions.
For more details, please refer to the \emph{Exploratory data analyses and preparation report}
(*** add link?). When it's done processing, the function exports the updated dataset
(\code{ndata_clean}) as a new .csv file.
}
\examples{
\dontrun{
mydata <- tdata_upD_final()$clean_dataset
mydata <- tdata_upD_final()$priorimp_dataset # May be useful to compare the results of missing
# data imputation.
myerror_pm <- tdata_upD_final()$impute_error_pm
myerror_cc <- tdata_upD_final()$impute_error_cc
path_to_csv <- tdata_upD_parcond()$path
}
}
