% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_01_prepare_data.R
\name{tdata_upD_final}
\alias{tdata_upD_final}
\title{Update and export the tits dataset completed with imputed missing values}
\usage{
tdata_upD_final(my_tdata = here::here("output", "tables", "ndata_parcond.csv"))
}
\arguments{
\item{my_tdata}{The name of the \strong{target} that contains the dataset generated
by \code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}. Note that this argument is not
necessarily required but is strongly advised.
Therefore, if you want to reproduce this work, you should not call this function unless you
have previously built the {targets} pipeline described in _targets.R (e.g. by using the
make.R file of the project)! If you DO NOT KNOW what I'm talking about, then you should
read the README file (https://github.com/mrelnoob/ppl.tits) and probably NOT use this function!}
}
\value{
A list containing 1) the updated dataset (a tibble accessible with
\code{tdata_upD_final()$dataset}), 2 and 3) tables showing the out-of-bag imputation error for both
species (accessible with \code{tdata_upD_final()$impute_error_pm} or
\code{tdata_upD_final()$impute_error_cc}), and 4) the path to the exported .csv file of the updated
dataset (accessible with \code{tdata_upD_final()$path}).
}
\description{
The function is the \strong{fourth} and last of a series of functions meant to
update and complete the tits dataset with its independent variables (i.e. predictors and
covariates).
The \code{tdata_upD_final} function loads and modifies the dataset generated by the previous
update function (\code{\link[ppl.tits:tdata_upD_parcond]{tdata_upD_parcond}}) by imputing missing
values using \strong{Random Forest} regressions. For more details, please refer to the
\emph{Exploratory data analyses and preparation report} (*** add link?).
When it's done processing, the function exports the updated dataset as a new .csv file.
}
\examples{
\dontrun{
mydata <- tdata_upD_final()$dataset
myerror_pm <- tdata_upD_final()$impute_error_pm
myerror_cc <- tdata_upD_final()$impute_error_cc
path_to_csv <- tdata_upD_parcond()$path
}
}
